requirements.txt
fastapi==0.111.0
elasticsearch[async]==8.13.2
redis==5.0.4
pydantic-settings==2.1.0